Move intervention_dist function to Intervention.cpp
