+++ Pv_mod/Source.cpp
@@ -67,14 +67,8 @@
 
 #include <iostream>
 #include <fstream>
-#include <stdlib.h>
-#include <math.h>
-#include <string>
-#include <time.h>
+#include <cmath>
 #include "randlib.h"
-#include <omp.h>
-#include <algorithm>
-#include <regex>
 
 
 //////////////////////////////////////////////////////
@@ -151,10 +145,8 @@ void mosq_derivs(const double t, double(&yM)[N_spec][N_M_comp], double(&dyMdt)[N
 void mosq_rk4(const double t, const double t_step_mosq, double(&yM)[N_spec][N_M_comp], Params& theta, Population& POP);
 void mosquito_step(double t, Params& theta, Population& POP);
 void human_step(Params& theta, Population& POP);
-void intervention_dist(double t, Params& theta, Population& POP, Intervention& INTVEN);
 void POP_summary(Population& POP);
 void model_simulator(Params& theta, Population& POP, Intervention& INTVEN, Simulation& SIM);
-double phi_inv(double pp, double mu, double sigma);
 
 
 ////////////////////////////////////////////
@@ -1226,826 +1218,4 @@ void model_simulator(Params& theta, Population& POP, Intervention& INTVEN, Simul
 
     }
 
-}
-
-
-
-//////////////////////////////////////////////////////////////////////////////
-//                                                                          //
-//  2.7. Vector control distribution                                        //
-//                                                                          //
-//////////////////////////////////////////////////////////////////////////////
-
-
-void intervention_dist(double t, Params& theta, Population& POP, Intervention& INTVEN)
-{
-    double QQ;
-
-    bool BS_effective;
-    bool PQ_treat;
-    bool PQ_effective;
-
-    bool MSAT_pos;
-    bool SSAT_pos;
-
-    //////////////////////////////////////////////////////////
-    // Intervention 1: LLINS
-
-    for (size_t m = 0; m<INTVEN.LLIN_year.size(); m++)
-    {
-        if ((t > INTVEN.LLIN_year[m] - 0.5*t_step) &&
-            (t < INTVEN.LLIN_year[m] + 0.51*t_step))
-        {
-            cout << "LLIN distribution" << endl;
-
-            try 
-            {
-                QQ = phi_inv(INTVEN.LLIN_cover[m], 0.0, sqrt(1.0 + theta.sig_round_LLIN*theta.sig_round_LLIN));
-            }
-            catch (const char* e)
-            {
-                std::cerr << e << std::endl;
-                exit (1);
-            }
-
-            for (int n = 0; n<POP.N_pop; n++)
-            {
-                if (gennor(POP.people[n].zz_int[0], theta.sig_round_LLIN) < QQ)
-                {
-                    POP.people[n].LLIN = 1;
-                    POP.people[n].LLIN_age = 0.0;
-
-                    for (int g = 0; g < N_spec; g++)
-                    {
-                        POP.people[n].d_LLIN[g] = theta.d_LLIN_0[g];
-                        POP.people[n].r_LLIN[g] = theta.r_LLIN_0[g];
-                        POP.people[n].s_LLIN[g] = 1.0 - POP.people[n].d_LLIN[g] - POP.people[n].r_LLIN[g];
-                    }
-                }
-            }
-        }
-    }
-
-
-    //////////////////////////////////////////////////////////
-    // Intervention 2: IRS
-
-    for (size_t m = 0; m<INTVEN.IRS_year.size(); m++)
-    {
-        if ((t > INTVEN.IRS_year[m] - 0.5*t_step) &&
-            (t < INTVEN.IRS_year[m] + 0.51*t_step))
-        {
-            cout << "IRS distribution" << endl;
-
-            try 
-            {
-                QQ = phi_inv(INTVEN.IRS_cover[m], 0.0, sqrt(1.0 + theta.sig_round_IRS*theta.sig_round_IRS));
-            }
-            catch (const char* e)
-            {
-                std::cerr << e << std::endl;
-                exit (1);
-            }
-
-            for (int n = 0; n<POP.N_pop; n++)
-            {
-                if (gennor(POP.people[n].zz_int[1], theta.sig_round_IRS) < QQ)
-                {
-                    POP.people[n].IRS = 1;
-                    POP.people[n].IRS_age = 0.0;
-
-                    for (int g = 0; g < N_spec; g++)
-                    {
-                        POP.people[n].d_IRS[g] = theta.d_IRS_0[g];
-                        POP.people[n].r_IRS[g] = theta.r_IRS_0[g];
-                        POP.people[n].s_IRS[g] = 1.0 - POP.people[n].d_IRS[g] - POP.people[n].r_IRS[g];
-                    }
-                }
-            }
-        }
-    }
-
-
-    //////////////////////////////////////////////////////////
-    // Intervention 3: first-line treatment (blood-stage)
-
-    for (size_t m = 0; m<INTVEN.BS_treat_year_on.size(); m++)
-    {
-        if ((t > INTVEN.BS_treat_year_on[m] - 0.5*t_step) &&
-            (t < INTVEN.BS_treat_year_on[m] + 0.51*t_step))
-        {
-            cout << "New front-line BS treatment" << endl;
-
-            theta.BS_treat_BScover = INTVEN.BS_treat_BScover[m];
-            theta.BS_treat_BSeff   = INTVEN.BS_treat_BSeff[m];
-            theta.BS_treat_BSproph = INTVEN.BS_treat_BSproph[m];
-
-            theta.treat_BScover = theta.BS_treat_BScover;
-            theta.treat_BSeff   = theta.BS_treat_BSeff;
-            theta.treat_PQavail = 0.0;
-            theta.r_P           = 1.0 / theta.BS_treat_BSproph;
-        }
-    }
-
-
-    //////////////////////////////////////////////////////////
-    // Switching back to baseline.
-
-    for (size_t m = 0; m<INTVEN.BS_treat_year_on.size(); m++)
-    {
-        if ((t > INTVEN.BS_treat_year_off[m] - 0.5*t_step) &&
-            (t < INTVEN.BS_treat_year_off[m] + 0.51*t_step))
-        {
-            cout << "End of changing front-line BS treatment" << endl;
-
-            theta.treat_BScover = theta.BS_treat_BScover_base;
-            theta.treat_BSeff   = theta.BS_treat_BSeff_base;
-            theta.treat_PQavail = 0.0;
-            theta.r_P           = 1.0 / theta.BS_treat_BSproph_base;
-        }
-    }
-
-
-    //////////////////////////////////////////////////////////
-    // Intervention 4: first-line treatment (primaquine)
-
-    for (size_t m = 0; m<INTVEN.PQ_treat_year_on.size(); m++)
-    {
-        if ((t > INTVEN.PQ_treat_year_on[m] - 0.5*t_step) &&
-            (t < INTVEN.PQ_treat_year_on[m] + 0.51*t_step))
-        {
-            cout << "New front-line PQ treatment" << endl;
-
-            theta.PQ_treat_BScover     = INTVEN.PQ_treat_BScover[m];
-            theta.PQ_treat_BSeff       = INTVEN.PQ_treat_BSeff[m];
-            theta.PQ_treat_BSproph     = INTVEN.PQ_treat_BSproph[m];
-            theta.PQ_treat_PQavail     = INTVEN.PQ_treat_PQavail[m];
-            theta.PQ_treat_PQeff       = INTVEN.PQ_treat_PQeff[m];
-            theta.PQ_treat_PQproph     = INTVEN.PQ_treat_PQproph[m];
-            theta.PQ_treat_G6PD_risk   = INTVEN.PQ_treat_G6PD_risk[m];
-            theta.PQ_treat_CYP2D6_risk = INTVEN.PQ_treat_CYP2D6_risk[m];
-            theta.PQ_treat_preg_risk   = INTVEN.PQ_treat_preg_risk[m];
-            theta.PQ_treat_low_age     = INTVEN.PQ_treat_low_age[m];
-
-            theta.treat_BScover = theta.PQ_treat_BScover;
-            theta.treat_BSeff   = theta.PQ_treat_BSeff;
-            theta.treat_PQavail = theta.PQ_treat_PQavail;
-            theta.r_P           = 1.0 / theta.PQ_treat_BSproph;
-        }
-    }
-
-
-    //////////////////////////////////////////////////////////
-    // Switching back to baseline.
-
-    for (size_t m = 0; m<INTVEN.PQ_treat_year_on.size(); m++)
-    {
-        if ((t > INTVEN.PQ_treat_year_off[m] - 0.5*t_step) &&
-            (t < INTVEN.PQ_treat_year_off[m] + 0.51*t_step))
-        {
-            cout << "End of changing front-line PQ treatment" << endl;
-
-            theta.PQ_treat_BScover     = 0.0;
-            theta.PQ_treat_BSeff       = 0.0;
-            theta.PQ_treat_BSproph     = 10.0;
-            theta.PQ_treat_PQavail     = 0.0;
-            theta.PQ_treat_PQeff       = 0.0;
-            theta.PQ_treat_PQproph     = 10.0;
-            theta.PQ_treat_G6PD_risk   = 1;
-            theta.PQ_treat_CYP2D6_risk = 1;
-            theta.PQ_treat_preg_risk   = 1;
-            theta.PQ_treat_low_age     = 180.0;
-
-            theta.treat_BScover = theta.BS_treat_BScover_base;
-            theta.treat_BSeff   = theta.BS_treat_BSeff_base;
-            theta.treat_PQavail = 0.0;
-            theta.r_P           = 1.0 / theta.BS_treat_BSproph_base;
-        }
-    }
-
-
-    //////////////////////////////////////////////////////////
-    // Intervention 5: MDA (blood-stage)
-
-    for (size_t m = 0; m<INTVEN.MDA_BS_year.size(); m++)
-    {
-        if ((t > INTVEN.MDA_BS_year[m] - 0.5*t_step) &&
-            (t < INTVEN.MDA_BS_year[m] + 0.51*t_step))
-        {
-            cout << "MDA (BS) distribution" << endl;
-
-            theta.MDA_BS_BScover = INTVEN.MDA_BS_BScover[m];
-            theta.MDA_BS_BSeff   = INTVEN.MDA_BS_BSeff[m];
-            theta.MDA_BS_BSproph = INTVEN.MDA_BS_BSproph[m];
-
-            try 
-            {
-                QQ = phi_inv(theta.MDA_BS_BScover, 0.0, sqrt(1.0 + theta.sig_round_MDA*theta.sig_round_MDA));
-            }
-            catch (const char* e)
-            {
-                std::cerr << e << std::endl;
-                exit (1);
-            }
-
-            for (int n = 0; n<POP.N_pop; n++)
-            {
-                if (gennor(POP.people[n].zz_int[2], theta.sig_round_MDA) < QQ)
-                {
-                    POP.people[n].ACT_treat = 1;
-
-                    if (genunf(0.0, 1.0) < theta.MDA_BS_BSeff)
-                    {
-                        if (POP.people[n].S == 1    ) { POP.people[n].S = 0;     POP.people[n].P = 1; }
-                        if (POP.people[n].I_PCR == 1) { POP.people[n].I_PCR = 0; POP.people[n].P = 1; }
-                        if (POP.people[n].I_LM == 1 ) { POP.people[n].I_LM = 0;  POP.people[n].P = 1; }
-                        if (POP.people[n].I_D == 1  ) { POP.people[n].I_D = 0;   POP.people[n].T = 1; }
-                    }
-                }
-            }
-        }
-    }
-
-
-    //////////////////////////////////////////////////////////
-    // Intervention 6: MDA (blood-stage and liver-stage)
-
-    for (size_t m = 0; m<INTVEN.MDA_PQ_year.size(); m++)
-    {
-        if ((t > INTVEN.MDA_PQ_year[m] - 0.5*t_step) &&
-            (t < INTVEN.MDA_PQ_year[m] + 0.51*t_step))
-        {
-            cout << "MDA (BS+PQ) distribution" << endl;
-
-            theta.MDA_PQ_BScover     = INTVEN.MDA_PQ_BScover[m];
-            theta.MDA_PQ_BSeff       = INTVEN.MDA_PQ_BSeff[m];
-            theta.MDA_PQ_BSproph     = INTVEN.MDA_PQ_BSproph[m];
-            theta.MDA_PQ_PQavail     = INTVEN.MDA_PQ_PQavail[m];
-            theta.MDA_PQ_PQeff       = INTVEN.MDA_PQ_PQeff[m];
-            theta.MDA_PQ_PQproph     = INTVEN.MDA_PQ_PQproph[m];
-            theta.MDA_PQ_G6PD_risk   = INTVEN.MDA_PQ_G6PD_risk[m];
-            theta.MDA_PQ_CYP2D6_risk = INTVEN.MDA_PQ_CYP2D6_risk[m];
-            theta.MDA_PQ_preg_risk   = INTVEN.MDA_PQ_preg_risk[m];
-            theta.MDA_PQ_low_age     = INTVEN.MDA_PQ_low_age[m];
-
-            try 
-            {
-                QQ = phi_inv(theta.MDA_PQ_BScover, 0.0, sqrt(1.0 + theta.sig_round_MDA*theta.sig_round_MDA));
-            }
-            catch (const char* e)
-            {
-                std::cerr << e << std::endl;
-                exit (1);
-            }
-
-            for (int n = 0; n<POP.N_pop; n++)
-            {
-                if (gennor(POP.people[n].zz_int[3], theta.sig_round_MDA) < QQ)
-                {
-                    /////////////////////////////////////////////////////
-                    // Blood-stage treatment is always administered
-                    // Is blood-stage treatment effective
-
-                    BS_effective = 0;
-
-                    if (genunf(0.0, 1.0) < theta.MDA_PQ_BSeff)
-                    {
-                        BS_effective = 1;
-                    }
-
-
-                    /////////////////////////////////////////////////////////////////////
-                    // Is PQ administered?
-
-                    PQ_treat = 0;
-
-                    if( genunf(0.0, 1.0) < theta.MDA_PQ_PQavail )
-                    {
-                        PQ_treat = 1;
-                    }
-                    
-
-                    /////////////////////////////////////////////////////////////////////
-                    // Exclude PQ because of G6PD deficiency
-
-                    if( (theta.MDA_PQ_G6PD_risk == 1) && (POP.people[n].G6PD_def == 1) )
-                    {
-                        PQ_treat = 0;
-                    }
-
-
-                    /////////////////////////////////////////////////////////////////////
-                    // Exclude PQ because of pregancy
-
-                    if( (theta.MDA_PQ_preg_risk == 1) && (POP.people[n].pregnant == 1) )
-                    {
-                        PQ_treat = 0;
-                    }
-
-
-                    /////////////////////////////////////////////////////////////////////
-                    // Exclude PQ because of young age
-
-                    if (POP.people[n].age < theta.MDA_PQ_low_age)
-                    {
-                        PQ_treat = 0;
-                    }
-
-                    if (PQ_treat == 1)
-                    {
-                        POP.people[n].PQ_treat = 1;
-                    }
-
-                    /////////////////////////////////////////////////////////////////////
-                    // Is PQ effective?
-
-                    PQ_effective = 1;
-
-                    if( genunf(0.0, 1.0) > theta.MDA_PQ_PQeff )
-                    {
-                        PQ_effective = 0;
-                    }
-
-                    if( (theta.MDA_PQ_CYP2D6_risk == 1) && (POP.people[n].CYP2D6 == 1) )
-                    {
-                        PQ_effective = 0;
-                    }
-
-
-                    /////////////////////////////////////////////////////////////////////
-                    // Was there PQ overtreatment?
-
-                    if( (PQ_treat == 1) && (POP.people[n].Hyp == 0) )
-                    {
-                        POP.people[n].PQ_overtreat = 1;
-                    }
-
-                    if( (PQ_treat == 1) && (POP.people[n].T_last_BS > 270.0) )
-                    {
-                        POP.people[n].PQ_overtreat_9m = 1;
-                    }
-
-
-                    /////////////////////////////////////////////////////////////////////
-                    // ACTION: administer blood-stage drug
-
-                    POP.people[n].ACT_treat = 1;
-
-                    if (BS_effective == 1)
-                    {
-                        if (POP.people[n].S == 1) {     POP.people[n].S = 0;     POP.people[n].P = 1; }
-                        if (POP.people[n].I_PCR == 1) { POP.people[n].I_PCR = 0; POP.people[n].P = 1; }
-                        if (POP.people[n].I_LM == 1) {  POP.people[n].I_LM = 0;  POP.people[n].P = 1; }
-                        if (POP.people[n].I_D == 1) {   POP.people[n].I_D = 0;   POP.people[n].T = 1; }
-                    }
-
-
-                    /////////////////////////////////////////////////////////////////////
-                    // ACTION: administer primaquine
-
-                    if ((PQ_treat == 1) && (PQ_effective == 1))
-                    {
-                        POP.people[n].Hyp = 0;
-
-                        POP.people[n].PQ_proph = 1;
-                        POP.people[n].PQ_proph_timer = theta.MDA_PQ_PQproph;
-                    }
-                }
-            }
-        }
-    }
-
-
-    //////////////////////////////////////////////////////////
-    // Intervention 7: MSAT (blood-stage and liver-stage)
-
-    for (size_t m = 0; m < INTVEN.MSAT_PQ_year.size(); m++)
-    {
-        if( (t > INTVEN.MSAT_PQ_year[m] - 0.5*t_step) &&
-            (t < INTVEN.MSAT_PQ_year[m] + 0.51*t_step) )
-        {
-            cout << "MSAT (BS+PQ) distribution" << endl;
-
-            theta.MSAT_PQ_BScover     = INTVEN.MSAT_PQ_BScover[m];
-            theta.MSAT_PQ_RDT_PCR     = INTVEN.MSAT_PQ_RDT_PCR[m];
-            theta.MSAT_PQ_sens        = INTVEN.MSAT_PQ_sens[m];
-            theta.MSAT_PQ_BSeff       = INTVEN.MSAT_PQ_BSeff[m];
-            theta.MSAT_PQ_BSproph     = INTVEN.MSAT_PQ_BSproph[m];
-            theta.MSAT_PQ_PQavail     = INTVEN.MSAT_PQ_PQavail[m];
-            theta.MSAT_PQ_PQeff       = INTVEN.MSAT_PQ_PQeff[m];
-            theta.MSAT_PQ_PQproph     = INTVEN.MSAT_PQ_PQproph[m];
-            theta.MSAT_PQ_G6PD_risk   = INTVEN.MSAT_PQ_G6PD_risk[m];
-            theta.MSAT_PQ_CYP2D6_risk = INTVEN.MSAT_PQ_CYP2D6_risk[m];
-            theta.MSAT_PQ_preg_risk   = INTVEN.MSAT_PQ_preg_risk[m];
-            theta.MSAT_PQ_low_age     = INTVEN.MSAT_PQ_low_age[m];
-
-            try 
-            {
-                QQ = phi_inv(theta.MSAT_PQ_BScover, 0.0, sqrt(1.0 + theta.sig_round_MDA*theta.sig_round_MDA));
-            }
-            catch (const char* e)
-            {
-                std::cerr << e << std::endl;
-                exit (1);
-            }
-
-            for (int n = 0; n < POP.N_pop; n++)
-            {
-                if (gennor(POP.people[n].zz_int[4], theta.sig_round_MDA) < QQ)
-                {
-                    /////////////////////////////////////////////////////
-                    // Blood-stage treatment is always administered
-
-                    MSAT_pos = 0;
-
-                    ////////////////////////////////////////////////
-                    // Diagnosis by RDT, assumed the same as LM 
-
-                    if (theta.MSAT_PQ_RDT_PCR == 1)
-                    {
-                        if ((POP.people[n].I_LM == 1) || (POP.people[n].I_D == 1) || (POP.people[n].T == 1))
-                        {
-                            if (genunf(0.0, 1.0) < theta.MSAT_PQ_sens)
-                            {
-                                MSAT_pos = 1;
-                            }
-                        }
-                    }
-
-
-                    ////////////////////////////////////////////////
-                    // Diagnosis by PCR 
-
-                    if (theta.MSAT_PQ_RDT_PCR == 2)
-                    {
-                        if ((POP.people[n].I_PCR == 1) || (POP.people[n].I_LM == 1) || (POP.people[n].I_D == 1) || (POP.people[n].T == 1))
-                        {
-                            if (genunf(0.0, 1.0) < theta.MSAT_PQ_sens)
-                            {
-                                MSAT_pos = 1;
-                            }
-                        }
-                    }
-
-
-                    /////////////////////////////////////////////////////
-                    // Is blood-stage treatment effective
-
-                    BS_effective = 0;
-
-                    if (genunf(0.0, 1.0) < theta.MSAT_PQ_BSeff)
-                    {
-                        BS_effective = 1;
-                    }
-
-
-                    /////////////////////////////////////////////////////////////////////
-                    // Is PQ administered?
-
-                    PQ_treat = 0;
-
-                    if (MSAT_pos == 1)
-                    {
-                        if (genunf(0.0, 1.0) < theta.MSAT_PQ_PQavail)
-                        {
-                            PQ_treat = 1;
-                        }
-                    }
-
-
-                    /////////////////////////////////////////////////////////////////////
-                    // Exclude PQ because of G6PD deficiency
-
-                    if ( (theta.MSAT_PQ_G6PD_risk == 1) && (POP.people[n].G6PD_def == 1) )
-                    {
-                        PQ_treat = 0;
-                    }
-
-                    /////////////////////////////////////////////////////////////////////
-                    // Exclude PQ because of pregancy
-
-                    if ( (theta.MSAT_PQ_preg_risk == 1) && (POP.people[n].pregnant == 1) )
-                    {
-                        PQ_treat = 0;
-                    }
-
-                    /////////////////////////////////////////////////////////////////////
-                    // Exclude PQ because of young age
-
-                    if (POP.people[n].age < theta.MSAT_PQ_low_age)
-                    {
-                        PQ_treat = 0;
-                    }
-
-                    if (PQ_treat == 1)
-                    {
-                        POP.people[n].PQ_treat = 1;
-                    }
-
-                    /////////////////////////////////////////////////////////////////////
-                    // Is PQ effective?
-
-                    PQ_effective = 1;
-
-                    if (genunf(0.0, 1.0) > theta.MSAT_PQ_PQeff)
-                    {
-                        PQ_effective = 0;
-                    }
-
-                    if ((theta.MSAT_PQ_CYP2D6_risk == 1) && (POP.people[n].CYP2D6 == 1))
-                    {
-                        PQ_effective = 0;
-                    }
-
-
-                    /////////////////////////////////////////////////////////////////////
-                    // Was there PQ overtreatment?
-
-                    if ((PQ_treat == 1) && (POP.people[n].Hyp == 0))
-                    {
-                        POP.people[n].PQ_overtreat = 1;
-                    }
-
-                    if ((PQ_treat == 1) && (POP.people[n].T_last_BS > 270.0))
-                    {
-                        POP.people[n].PQ_overtreat_9m = 1;
-                    }
-
-
-                    /////////////////////////////////////////////////////////////////////
-                    // ACTION: administer blood-stage drug
-
-                    if (MSAT_pos == 1)
-                    {
-                        POP.people[n].ACT_treat = 1;
-
-                        if (BS_effective == 1)
-                        {
-                            if (POP.people[n].S == 1) {     POP.people[n].S = 0;     POP.people[n].P = 1; }
-                            if (POP.people[n].I_PCR == 1) { POP.people[n].I_PCR = 0; POP.people[n].P = 1; }
-                            if (POP.people[n].I_LM == 1) {  POP.people[n].I_LM = 0;  POP.people[n].P = 1; }
-                            if (POP.people[n].I_D == 1) {   POP.people[n].I_D = 0;   POP.people[n].T = 1; }
-                        }
-                    }
-
-                    
-                    /////////////////////////////////////////////////////////////////////
-                    // ACTION: administer primaquine
-
-                    if ((PQ_treat == 1) && (PQ_effective == 1))
-                    {
-                        POP.people[n].Hyp = 0;
-
-                        POP.people[n].PQ_proph = 1;
-                        POP.people[n].PQ_proph_timer = theta.MSAT_PQ_PQproph;
-                    }
-                }
-            }
-        }
-    }
-
-
-    //////////////////////////////////////////////////////////
-    // Intervention 8: SSAT (blood-stage and liver-stage)
-
-    for (size_t m = 0; m<INTVEN.SSAT_PQ_year.size(); m++)
-    {
-        if ((t > INTVEN.SSAT_PQ_year[m] - 0.5*t_step) &&
-            (t < INTVEN.SSAT_PQ_year[m] + 0.51*t_step))
-        {
-            cout << "SSAT (BS+PQ) distribution" << endl;
-
-            theta.SSAT_PQ_BScover     = INTVEN.SSAT_PQ_BScover[m];
-            theta.SSAT_PQ_sens        = INTVEN.SSAT_PQ_sens[m];
-            theta.SSAT_PQ_spec        = INTVEN.SSAT_PQ_spec[m];
-            theta.SSAT_PQ_BSeff       = INTVEN.SSAT_PQ_BSeff[m];
-            theta.SSAT_PQ_BSproph     = INTVEN.SSAT_PQ_BSproph[m];
-            theta.SSAT_PQ_PQavail     = INTVEN.SSAT_PQ_PQavail[m];
-            theta.SSAT_PQ_PQeff       = INTVEN.SSAT_PQ_PQeff[m];
-            theta.SSAT_PQ_PQproph     = INTVEN.SSAT_PQ_PQproph[m];
-            theta.SSAT_PQ_G6PD_risk   = INTVEN.SSAT_PQ_G6PD_risk[m];
-            theta.SSAT_PQ_CYP2D6_risk = INTVEN.SSAT_PQ_CYP2D6_risk[m];
-            theta.SSAT_PQ_preg_risk   = INTVEN.SSAT_PQ_preg_risk[m];
-            theta.SSAT_PQ_low_age     = INTVEN.SSAT_PQ_low_age[m];
-
-            try 
-            {
-                QQ = phi_inv(theta.SSAT_PQ_BScover, 0.0, sqrt(1.0 + theta.sig_round_MDA*theta.sig_round_MDA));
-            }
-            catch (const char* e)
-            {
-                std::cerr << e << std::endl;
-                exit (1);
-            }
-
-            for (int n = 0; n < POP.N_pop; n++)
-            {
-                if (gennor(POP.people[n].zz_int[5], theta.sig_round_MDA) < QQ)
-                {
-                    /////////////////////////////////////////////////////
-                    // Blood-stage treatment is always administered
-
-                    POP.people[n].ACT_treat = 1;
-
-
-                    /////////////////////////////////////////////////////
-                    // Is blood-stage treatment effective
-
-                    BS_effective = 0;
-
-                    if( genunf(0.0, 1.0) < theta.SSAT_PQ_BSeff )
-                    {
-                        BS_effective = 1;
-                    }
-
-                    /////////////////////////////////////////////////////////////////////
-                    // SSAT screening for blood-stage infection in the last 9 months
-                    //
-                    // There are two options here. Option 1 define over-treatment on the 
-                    // basis of blood-stage infection with the last 9 month. Option 2
-                    // defines over-treatment on the basis of presence of hypnozoites.
-
-                    SSAT_pos = 0;
-
-                    // OPTION 1
-/*
-                    if( (POP.people[n].T_last_BS <= 270.0) && (genunf(0.0, 1.0) < theta.SSAT_PQ_sens) )
-                    {
-                        SSAT_pos = 1;
-                    }
-
-                    if( (POP.people[n].T_last_BS > 270.0) && (genunf(0.0, 1.0) > theta.SSAT_PQ_spec) )
-                    {
-                        SSAT_pos = 1;
-                    }
-*/
-
-                    // OPTION 2
-
-                    if ((POP.people[n].Hyp > 0) && (genunf(0.0, 1.0) < theta.SSAT_PQ_sens))
-                    {
-                        SSAT_pos = 1;
-                    }
-
-                    if ((POP.people[n].Hyp == 0) && (genunf(0.0, 1.0) > theta.SSAT_PQ_spec))
-                    {
-                        SSAT_pos = 1;
-                    }
-
-
-                    /////////////////////////////////////////////////////////////////////
-                    // Is PQ administered?
-
-                    PQ_treat = 0;
-
-                    if( SSAT_pos == 1 )
-                    {
-                        if( genunf(0.0, 1.0) < theta.SSAT_PQ_PQavail )
-                        {
-                            PQ_treat = 1;
-                        }
-                    }
-
-
-                    /////////////////////////////////////////////////////////////////////
-                    // Exclude PQ because of G6PD deficiency
-
-                    if( (theta.SSAT_PQ_G6PD_risk == 1) && (POP.people[n].G6PD_def == 1) )
-                    {
-                        PQ_treat = 0;
-                    }
-
-                    /////////////////////////////////////////////////////////////////////
-                    // Exclude PQ because of pregancy
-
-                    if( (theta.SSAT_PQ_preg_risk == 1) && (POP.people[n].pregnant == 1) )
-                    {
-                        PQ_treat = 0;
-                    }
-
-                    /////////////////////////////////////////////////////////////////////
-                    // Exclude PQ because of young age
-
-                    if( POP.people[n].age < theta.SSAT_PQ_low_age )
-                    {
-                        PQ_treat = 0;
-                    }
-
-                    if( PQ_treat == 1 )
-                    {
-                        POP.people[n].PQ_treat = 1;
-                    }
-
-                    /////////////////////////////////////////////////////////////////////
-                    // Is PQ effective?
-
-                    PQ_effective = 0;
-
-                    if( genunf(0.0, 1.0) < theta.SSAT_PQ_PQeff )
-                    {
-                        PQ_effective = 1;
-                    }
-
-                    if( (theta.SSAT_PQ_CYP2D6_risk == 1) && (POP.people[n].CYP2D6 == 1) )
-                    {
-                        PQ_effective = 0;
-                    }
-
-
-                    /////////////////////////////////////////////////////////////////////
-                    // Was there PQ overtreatment?
-
-                    if ((PQ_treat == 1) && (POP.people[n].Hyp == 0))
-                    {
-                        POP.people[n].PQ_overtreat = 1;
-                    }
-
-                    if ((PQ_treat == 1) && (POP.people[n].T_last_BS > 270.0))
-                    {
-                        POP.people[n].PQ_overtreat_9m = 1;
-                    }
-
-
-                    /////////////////////////////////////////////////////////////////////
-                    // ACTION: administer blood-stage drug
-
-                    if( BS_effective == 1 )
-                    {
-                        if (POP.people[n].S     == 1) { POP.people[n].S = 0;     POP.people[n].P = 1; }
-                        if (POP.people[n].I_PCR == 1) { POP.people[n].I_PCR = 0; POP.people[n].P = 1; }
-                        if (POP.people[n].I_LM  == 1) { POP.people[n].I_LM = 0;  POP.people[n].P = 1; }
-                        if (POP.people[n].I_D   == 1) { POP.people[n].I_D = 0;   POP.people[n].T = 1; }
-                    }
-
-                    /////////////////////////////////////////////////////////////////////
-                    // ACTION: administer primaquine
-
-                    if( (PQ_treat == 1) && (PQ_effective == 1) )
-                    {
-                        POP.people[n].Hyp = 0;
-
-                        POP.people[n].PQ_proph = 1;
-                        POP.people[n].PQ_proph_timer = theta.SSAT_PQ_PQproph;
-                    }
-                }
-            }
-
-
-        }
-    }
-
-}
-
-
-//////////////////////////////////////////////////////////////////////////////////////////
-////////////////////////////////////////////////////////////////////////////////////////// 
-//        //                                                                            // 
-//  2.9.  //  Inverse of the cumulative normal distribution function required           //
-//        //  for implementing correlated intervention coverage.                        //
-////////////                                                                            //
-////////////  The code is based on the following website                                //
-////////////  http://www.johndcook.com/blog/cpp_phi_inverse/                            //
-////////////  which is based the algorithm in Abramowitz and Stegun formula 26.2.23.    // 
-////////////                                                                            //
-////////////  The absolute value of the error should be less than 4.5 e-4 and it tests  //
-////////////  out nicely in R.                                                          //
-////////////                                                                            //
-//////////////////////////////////////////////////////////////////////////////////////////
-//////////////////////////////////////////////////////////////////////////////////////////
-
-double phi_inv(double pp, double mu, double sigma)
-{
-    if (pp < 0.0 || pp > 1.0)
-    {
-        throw("bad vlaue of pp (coverage) in phi_inv");
-    }
-    else if (pp == 0.0) 
-    {
-        return -std::numeric_limits<double>::infinity();
-    }
-    else if (pp == 1.0)
-    {
-        return std::numeric_limits<double>::infinity();
-    }
-
-    double cc[] = { 2.515517, 0.802853, 0.010328 };
-    double dd[] = { 1.432788, 0.189269, 0.001308 };
-
-    double tt, temp;
-
-    if (pp < 0.5)
-    {
-        tt = sqrt(-2.0*log(pp));
-
-        temp = -(tt - ((cc[2] * tt + cc[1])*tt + cc[0]) / (((dd[2] * tt + dd[1])*tt + dd[0])*tt + 1.0));
-
-    }
-    else {
-
-        tt = sqrt(-2.0*log(1.0 - pp));
-
-        temp = (tt - ((cc[2] * tt + cc[1])*tt + cc[0]) / (((dd[2] * tt + dd[1])*tt + dd[0])*tt + 1.0));
-    }
-
-    return mu + sigma*temp;
 }
